wisteria-web/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
openai==1.3.0
PyYAML==6.0.1
backoff==2.2.1
python-socketio==5.9.0
python-engineio==4.7.1
Werkzeug==2.3.7
SQLAlchemy>=2.0.30
python-dotenv==1.0.0 